Move ev charger power and current formulas to EVChargerPool
... from the logical meter
