+++ src/frequenz/sdk/timeseries/ev_charger_pool/_ev_charger_pool.py
@@ -6,11 +6,18 @@
 from __future__ import annotations
 
 import logging
+import uuid
 
-from frequenz.channels import Receiver
+from frequenz.channels import Receiver, Sender
 
 from ... import microgrid
+from ...actor import ChannelRegistry, ComponentMetricRequest
 from ...microgrid.component import ComponentCategory
+from .._formula_engine import FormulaEnginePool, FormulaReceiver, FormulaReceiver3Phase
+from .._formula_engine._formula_generators import (
+    EVChargerCurrentFormula,
+    EVChargerPowerFormula,
+)
 from ._state_tracker import EVChargerPoolStates, StateTracker
 
 logger = logging.getLogger(__name__)
@@ -21,16 +28,24 @@ class EVChargerPool:
 
     def __init__(
         self,
+        channel_registry: ChannelRegistry,
+        resampler_subscription_sender: Sender[ComponentMetricRequest],
         component_ids: set[int] | None = None,
     ) -> None:
         """Create an `EVChargerPool` instance.
 
         Args:
+            channel_registry: A channel registry instance shared with the resampling
+                actor.
+            resampler_subscription_sender: A sender for sending metric requests to the
+                resampling actor.
             component_ids: An optional list of component_ids belonging to this pool.  If
                 not specified, IDs of all ev chargers in the microgrid will be fetched
                 from the component graph.
         """
-        self._component_ids = set()
+        self._channel_registry = channel_registry
+        self._resampler_subscription_sender = resampler_subscription_sender
+        self._component_ids: set[int] = set()
         if component_ids is not None:
             self._component_ids = component_ids
         else:
@@ -42,6 +57,12 @@ class EVChargerPool:
                 )
             }
         self._state_tracker: StateTracker | None = None
+        self._namespace = f"ev-charger-pool-{uuid.uuid4()}"
+        self._formula_pool = FormulaEnginePool(
+            self._namespace,
+            self._channel_registry,
+            self._resampler_subscription_sender,
+        )
 
     async def _stop(self) -> None:
         if self._state_tracker:
@@ -57,3 +78,31 @@ class EVChargerPool:
         if not self._state_tracker:
             self._state_tracker = StateTracker(self._component_ids)
         return self._state_tracker.new_receiver()
+
+    async def total_current(self) -> FormulaReceiver3Phase:
+        """Fetch the total current for the ev chargers in the pool.
+
+        If a formula engine to calculate ev charger current is not already
+        running, it will be started.  Else, we'll just get a new receiver to the
+        already existing data stream.
+
+        Returns:
+            A *new* receiver that will stream ev_charger current values.
+        """
+        return await self._formula_pool.from_generator(
+            "ev_charger_total_current", EVChargerCurrentFormula
+        )
+
+    async def total_power(self) -> FormulaReceiver:
+        """Fetch the total power for the ev chargers in the pool.
+
+        If a formula engine to calculate EV charger power is not already
+        running, it will be started. Else, we'll just get a new receiver
+        to the already existing data stream.
+
+        Returns:
+            A *new* receiver that will stream ev_charger power values.
+        """
+        return await self._formula_pool.from_generator(
+            "ev_charger_total_power", EVChargerPowerFormula
+        )
