+++ tests/timeseries/test_logical_meter.py
@@ -235,45 +235,3 @@ class TestLogicalMeter:
             assert isclose(soc_sample.value, sum(bat_vals) / len(bat_vals))
 
         await mockgrid.cleanup()
-
-    async def test_ev_power(  # pylint: disable=too-many-locals
-        self,
-        mocker: MockerFixture,
-    ) -> None:
-        """Test the battery power and pv power formulas."""
-        mockgrid = MockMicrogrid(grid_side_meter=False)
-        mockgrid.add_ev_chargers(5)
-        request_chan, channel_registry = await mockgrid.start(mocker)
-        logical_meter = LogicalMeter(
-            channel_registry,
-            request_chan.new_sender(),
-            microgrid.get().component_graph,
-        )
-
-        main_meter_recv = await get_resampled_stream(
-            logical_meter,
-            channel_registry,
-            request_chan.new_sender(),
-            mockgrid.main_meter_id,
-            ComponentMetricId.ACTIVE_POWER,
-        )
-        grid_power_recv = await logical_meter.grid_power()
-        ev_power_recv = await logical_meter.ev_charger_power()
-
-        await synchronize_receivers([grid_power_recv, main_meter_recv, ev_power_recv])
-
-        ev_results = []
-        for _ in range(10):
-            grid_pow = await grid_power_recv.receive()
-            ev_pow = await ev_power_recv.receive()
-            main_pow = await main_meter_recv.receive()
-
-            assert grid_pow is not None and grid_pow.value is not None
-            assert ev_pow is not None and ev_pow.value is not None
-            assert main_pow is not None and main_pow.value is not None
-            assert isclose(grid_pow.value, ev_pow.value + main_pow.value)
-
-            ev_results.append(ev_pow.value)
-
-        await mockgrid.cleanup()
-        assert len(ev_results) == 10
